test: combine a few more tests
