+++ Cargo.toml
@@ -6,7 +6,6 @@ members = [
   "prqlc/prql-compiler-macros",
   "prqlc/prqlc",
   "prqlc/prql-compiler/examples/compile-files", # An example
-  "prqlc/tests_misc",
   "prqlc/bindings/elixir/native/prql",
   "prqlc/bindings/java",
   "prqlc/bindings/js",
