+++ prqlc/prql-compiler/tests/sql/bad_error_messages.rs
