+++ prqlc/prqlc-parser/Cargo.toml
@@ -10,6 +10,7 @@ version.workspace = true
 
 [lib]
 doctest = false
+test = false
 
 [dependencies]
 itertools = "0.11.0"
