+++ prqlc/prqlc/Cargo.toml
@@ -8,6 +8,10 @@ repository.workspace = true
 rust-version = "1.70.0"
 version.workspace = true
 
+[[bin]]
+name = "prqlc"
+test = false
+
 [target.'cfg(not(target_family="wasm"))'.dependencies]
 anstream = {version = "0.3.2", features = ["auto"]}
 anyhow = {version = "1.0.75"}
