--- prqlc/tests_misc/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "misc"
-version = "0.10.0"
-edition = "2021"
-publish = false
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dev-dependencies]
-insta = {version = "1.34", features = ["colors", "glob", "yaml"]}
-regex = "1.10.2"
-similar = {version = "2.3.0"}
